% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableList.R
\name{variableList}
\alias{variableList}
\title{Get Variable's Names from WS2 and formate them}
\usage{
variableList(wsUrl = "www.opensilex.org/openSilexAPI/rest/")
}
\arguments{
\item{wsUrl}{url of the webservice}
}
\value{
WSResponse
}
\description{
Get Variable's Names from WS2 and formate them
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_2_public","guest@opensilex.org","guest")
 variableList()
}
}
